Refactor conflux_web3.contract.py to a submodule
